moved files from deployment-->k8sresources
