+++ test/e2e/k8sresources/rbac.go
@@ -14,76 +14,90 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package deployment
+package k8sresources
 
 import (
+	"context"
+
 	"github.com/google/uuid"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	log "github.com/sirupsen/logrus"
+
 	k8s "github.com/soda-cdm/kahu/test/e2e/util/k8s"
 	kahu "github.com/soda-cdm/kahu/test/e2e/util/kahu"
 )
 
-//testcase for E2E service backup and restore
-var _ = Describe("ServiceBackup", Label("service"), func() {
-	Context("Create backup of service and restore", func() {
-		It("service", func() {
+//testcase for E2E deployment backup and restore
+var _ = Describe("rbacBackup", Label("rbac"), func() {
+	Context("Create backup of rbac and restore", func() {
+		It("rbac", func() {
 			kubeClient, kahuClient := kahu.Clients()
-			//Create service to test
+			//Create rbac to test
 			ns := kahu.BackupNameSpace
+			ctx := context.TODO()
 			UUIDgen, err := uuid.NewRandom()
 			Expect(err).To(BeNil())
-			name := "service" + "-" + UUIDgen.String()
-			service, err := k8s.CreateService(kubeClient, ns, name)
-			log.Infof("service,err:%v,%v\n", service, err)
-			Expect(err).To(BeNil())
-			err = k8s.WaitUntilServiceCreated(kubeClient, ns, name)
+			clusterRolename := "clusterrole" + "-" + UUIDgen.String()
+			serviceAccount := "serviceaccount" + "-" + UUIDgen.String()
+			clusterrolebinding := "clusterrolebinding" + "-" + UUIDgen.String()
+			podName := UUIDgen.String() + "-" + "pod"
+
+			//create serviceAccount
+			err = k8s.CreateServiceAccount(ctx, kubeClient, ns, serviceAccount)
 			Expect(err).To(BeNil())
 
-			//create backup for the service
-			backupName := "backup" + "service" + "-" + UUIDgen.String()
+			//create pod with the serviceAccount
+			pod, err := k8s.CreatePodWithServiceAccount(kubeClient, ns, podName, serviceAccount)
+			log.Infof("pod with sa is %v\n", pod)
+			Expect(err).To(BeNil())
+
+			err = k8s.CreateRBACWithBindingSA(ctx, kubeClient, ns, serviceAccount, clusterRolename, clusterrolebinding)
+			Expect(err).To(BeNil())
+
+			_, err = k8s.GetClusterRoleBinding(ctx, kubeClient, clusterrolebinding)
+			Expect(err).To(BeNil())
+
+			_, err = k8s.GetClusterRole(ctx, kubeClient, clusterRolename)
+			Expect(err).To(BeNil())
+
+			//create backup for the rbac
+			backupName := "backup" + "rbac" + "-" + UUIDgen.String()
 			includeNs := kahu.BackupNameSpace
-			resourceType := "Service"
+			resourceType := "Pod"
 			_, err = kahu.CreateBackup(kahuClient, backupName, includeNs, resourceType)
 			Expect(err).To(BeNil())
 			err = kahu.WaitForBackupCreate(kahuClient, backupName)
 			Expect(err).To(BeNil())
-			log.Infof("backup of service is done %v\n", backupName)
+			log.Infof("backup of rbac is done\n")
 
 			// create restore for the backup
-			restoreName := "restore" + "service" + "-" + UUIDgen.String()
+			restoreName := "restore" + "rbac" + "-" + UUIDgen.String()
 			nsRestore := kahu.RestoreNameSpace
 			restore, err := kahu.CreateRestore(kahuClient, restoreName, backupName, includeNs, nsRestore)
 			log.Debugf("restore1 is %v\n", restore)
 			Expect(err).To(BeNil())
 			err = kahu.WaitForRestoreCreate(kahuClient, restoreName)
 			Expect(err).To(BeNil())
+			log.Infof("restore is created\n")
 
-			//check if the restored service is up
-			service, err = k8s.GetService(kubeClient, name, nsRestore)
-			log.Debugf("service is %v\n", service)
+			//check if the restored rbac is up
+			err = k8s.WaitForPodComplete(kubeClient, nsRestore, podName)
 			Expect(err).To(BeNil())
-
-			err = k8s.WaitUntilServiceCreated(kubeClient, nsRestore, name)
-			Expect(err).To(BeNil())
-			log.Infof("restored service %v is up\n", name)
 
 			//Delete the. restore
 			err = kahu.DeleteRestore(kahuClient, restoreName)
 			Expect(err).To(BeNil())
 			err = kahu.WaitForRestoreDelete(kahuClient, restoreName)
 			Expect(err).To(BeNil())
-			log.Infof("restore of %v is deleted\n", name)
+			log.Infof("restore %v is deleted\n", restoreName)
 
 			//Delete the backup
 			err = kahu.DeleteBackup(kahuClient, backupName)
 			Expect(err).To(BeNil())
 			err = kahu.WaitForBackupDelete(kahuClient, backupName)
 			Expect(err).To(BeNil())
-			log.Infof("backup of  %v is deleted\n", name)
+			log.Infof("backup %v is deleted\n", backupName)
 		})
 	})
 })
-
-//TODO need to create namespace include steps
